requirements.txt
wandb>=0.16.0
pynvml>=11.5.0
psutil>=5.9.0
APScheduler>=3.10.0
requests>=2.31.0
matplotlib>=3.7.0
openai>=1.0.0
anthropic>=0.20.0
zhipuai>=2.0.0
PyYAML>=6.0
